requirements.txt
# PDF Processing
pymupdf>=1.23.0
pymupdf-fonts>=1.0.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0
# For GPU support, install paddlepaddle-gpu instead:
# pip install paddlepaddle-gpu

# Translation
google-generativeai>=0.3.0

# Language Detection
langdetect>=1.0.9

# Web Frontend
streamlit>=1.29.0

# Environment
python-dotenv>=1.0.0

# Testing
hypothesis>=6.92.0
pytest>=7.4.0
